% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{clusterPlot}
\alias{clusterPlot}
\title{Plot a flow-based domain of a typical day}
\usage{
clusterPlot(data, country1, country2, hour, dayType,
  typicalDayOnly = FALSE, ggplot = FALSE, width = "420px",
  height = "410px", xlim = c(-10000, 10000), ylim = c(-10000, 10000),
  export = T)
}
\arguments{
\item{data}{\code{data.table} results from the clustering, output data from \link{clusterTypicalDaysForOneClass}}

\item{country1}{\code{character}, name of the country whose net position is in the x axis (BE, FR, DE or NL)}

\item{country2}{\code{character}, name of the country whose net position is in the y axis (BE, FR, DE or NL)}

\item{hour}{\code{numeric}, hour of the plotted domain}

\item{dayType}{\code{numeric}, typical flow-based day identifier}

\item{typicalDayOnly}{\code{logical} if TRUE, plot only the domain of the typical day 
and not the other domains of the cluster}

\item{ggplot}{\code{logical} should ggplot package be used (static graph) instead of rAmCharts (dynamic graph)}

\item{width}{\code{character}, for rAmCharts only. Default to "420px" (set to "100/100" for dynamic resize)}

\item{height}{\code{character}, for rAmCharts only. Default to "410px" (set to "100/100" for dynamic resize)}

\item{xlim}{\code{numeric}, limits of x-axis (default = c(-10000, 10000))}

\item{ylim}{\code{numeric}, limits of x-axis (default = c(-10000, 10000))}

\item{export}{\code{logical} If you want to have the possibility to export your
graphic (if true in rmarkdown, can return a blank html)}
}
\description{
For a given hour, plot the flow-based domain of one typical day along with the other
flow-based domains it represents (i.e. other domains of the same cluster).
}
\examples{

\dontrun{

# classification result
clusterTD <- readRDS(system.file("testdata/allTypeDaytest.rds",package = "fbClust"))

clusterPlot(clusterTD, "FR", "DE", 3, 1, FALSE, FALSE)
clusterPlot(clusterTD, "ptdfFR", "DE", 3, 1, FALSE, TRUE)
clusterPlot(clusterTD, "FR", "ptdfDE", 3, 1, TRUE, TRUE)
clusterPlot(clusterTD, "ptdfFR", "ptdfDE", 3, 1, TRUE, FALSE)

}

}
