% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerator.R
\name{generateClusteringReport}
\alias{generateClusteringReport}
\title{Generate html report}
\usage{
generateClusteringReport(dayType, outputFile, data,
  countries = list(c("BE", "FR"), c("BE", "NL"), c("DE", "FR"), c("DE",
  "AT")), xlim = c(-10000, 10000), ylim = c(-10000, 10000))
}
\arguments{
\item{dayType}{\code{numeric} Typical day identifier}

\item{outputFile}{\code{character} a folder where the html report is saved}

\item{data}{\code{data.table} results from the clustering, output data from \link{clusteringTypicalDays}}

\item{countries}{\code{list, character} a list of couples of countries to choose the axises for the projection
of the flowbased domains (ex : list(c("BE", "FR"), c("BE", "NL"))) or an array of countries
(ex : c("FR", "NL", "AT")) to project the domains on all the countries combination 
(here FR+NL, FR+AT and NL+AT)}

\item{xlim}{\code{numeric}, limits of x-axis (default = c(-10000, 10000))}

\item{ylim}{\code{numeric}, limits of x-axis (default = c(-10000, 10000))}
}
\description{
Write and save a report with the results of the clustering of the flow-based domains. For one typical day,
the report contains the representation of all the 24 hourly flow-based domains and their comparison
with the domains of the other historical days it represents.
}
\examples{

\dontrun{

allTypDay <- readRDS(system.file("testdata/allTypDays.rds", package = "fbClust"))
generateClusteringReport(dayType = 1, data = allTypDay, outputFile = tempdir(),
countries = list(c("BE", "FR"), c("BE", "NL")), xlim = c(-12000, 12000),
ylim = c(-12000, 12000))

generateClusteringReport(dayType = 1, data = allTypDay, outputFile = tempdir(),
countries = c("AT", "FR", "NL", "DE"), xlim = c(-12000, 12000),
ylim = c(-12000, 12000))
}
}
