% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVertices.R
\name{getVertices}
\alias{getVertices}
\title{Give the vertices from PTDF data}
\usage{
getVertices(PTDF, ctrdel = NULL)
}
\arguments{
\item{PTDF}{{data.frame | data.table} PLAN, at least Date, Period and two ptdf columns :

\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
 \item ram : line limits
 \item Date : date in format YYYY-MM-DD
 \item Period : hour in the day, between 1 and 24
}}

\item{ctrdel}{{character} name of country deleted (two maj letters,
ex : FR for France or NL for Nederlands), can be NULL}
}
\description{
This function generates the vertices of the polyhedra from the facets
equation. It can only be computed after using \link{setDiffNotWantedPtdf} which
itself compute the plans equations.
Otherwise, it will crash because the ptdf won't correspond to facets equations.
}
\examples{
\dontrun{
PLAN <- readRDS(system.file("testdata/plan_test.rds", package = "fbClust"))

VERT <- getVertices(PLAN)
}
}
