% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plotFlowbased}
\alias{plotFlowbased}
\title{Plot flow-based domain(s)}
\usage{
plotFlowbased(PLAN, country1, country2, hours, dates,
  domainsNames = NULL, country_list = list(NL = c("BE", "DE", "FR",
  "AT")), xlim = c(-10000, 10000), ylim = c(-10000, 10000),
  main = NULL, width = "420px", height = "410px")
}
\arguments{
\item{PLAN}{\code{data.table}, at least ram, Date, Period and two ptdf columns :
\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
 \item ram : line limits
 \item Date : date in format YYYY-MM-DD
 \item Period : hour in the day, between 1 and 24
}
PLAN is generated in this format with the function \link{getPreprocPlan}}

\item{country1}{\code{character}, name of the country whose net position is in the x axis (BE, FR, DE, AT or NL)}

\item{country2}{\code{character}, name of the country whose net position is in the y axis (BE, FR, DE, AT or NL)}

\item{hours}{\code{character}, hours of interest for the graphics.}

\item{dates}{\code{character}, dates of interest for the graphics. (All the combinations of dates and hours are )}

\item{domainsNames}{\code{character} names of the domain(s), used as legend of the graphics.
The length of domainsNames has to be the same of the number of combinations of hours and dates}

\item{country_list}{\code{list}, list of countries in the ptdf, with the ones which should
sustracted to the others as the names of the arrays which themself contain the ones which
be sustracted}

\item{xlim}{\code{numeric}, limits of x-axis (default = c(-10000, 10000))}

\item{ylim}{\code{numeric}, limits of x-axis (default = c(-10000, 10000))}

\item{main}{\code{character} title of the graph, if NULL, the title will be "Domains country1 - country2"}

\item{width}{\code{character}, for rAmCharts only. Default to "420px" (set to "100/100" for dynamic resize)}

\item{height}{\code{character}, for rAmCharts only. Default to "410px" (set to "100/100" for dynamic resize)}
}
\description{
Plot flow-based domain(s)
}
\examples{

\dontrun{
library(data.table)
library(rAmCharts)
PLAN <- getPreprocPlan(
path_ptdf_matrix_factor = system.file(
  "testdata/plan_new_version_factor_AT.rds", package = "fbClust"),
path_ptdf_matrix_constraint = system.file(
 "testdata/plan_new_version_constraint_AT.rds", package = "fbClust"))
 
 country_list = list(NL = c("BE", "DE", "FR", "AT"))
#Plot unique polyhedron
plotFlowbased(PLAN, "BE", "DE", country_list = country_list, 
hours = c(2), dates = c("2018-10-02"), domainsNames = "2018-10-02", main = "")

#Plot four polyhedra
plotFlowbased(PLAN, "BE", "DE", country_list = country_list, 
hours = c(3, 4), dates = c("2018-10-02", "2018-10-04"), domainsNames = NULL,
main = NULL)

}

}
