% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDiffNotWantedPtdf.R
\name{setDiffNotWantedPtdf}
\alias{setDiffNotWantedPtdf}
\title{Compute differance for remove contry}
\usage{
setDiffNotWantedPtdf(PLAN, not_wanted_col = NULL)
}
\arguments{
\item{PLAN}{{data.table}}

\item{not_wanted_col}{{character}}
}
\description{
Remove country in not_wanted_col
}
