% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateAllTypeDays.R
\name{writeAllTypeDays}
\alias{writeAllTypeDays}
\title{Write the output data of a clustering in csv file}
\usage{
writeAllTypeDays(allTypeDay, output = "ptdf", outputFile, csv = T,
  onlyTypicalDay = T)
}
\arguments{
\item{allTypeDay}{\code{data.table} the output data.table of one of the functions 
\link{clusterTypicalDaysForOneClass} and \link{clusteringTypicalDays}.
\itemize{
 \item TypicalDay : The typical day of the cluster defined by the idDayType column.
 \item Class : The class of the clustering (typically, one of the classes obtained
 with the \link{getCalendar} function).
 \item dayIn : The details of the days in the cluster.
 \item distance : The distance of each day of the cluster to the typical day.
 \item idDayType : The id of the cluster.
}}

\item{output}{\code{character} The type of output you want to return 
(either ptdf, ptdfraw, vertices), default ptdf
\itemize{
 \item ptdf : The ptdf obtained after differenciation 
 \item ptdfraw : The raw ptdf (obtained just after \link{getPreprocPlan})
 \item vertices : The vertices obtained with the function \link{getVertices}
}}

\item{outputFile}{\code{character} The directory where you want to save the file.}

\item{csv}{\code{logical} If you want to write a csv or a rds, default TRUE}

\item{onlyTypicalDay}{\code{logical} If you want only the typical days or all 
days of the clustering}
}
\description{
This function has been built to give the possibility to the user
to keep the results of the clustering in a csv file using the output object of 
the functions \link{clusterTypicalDaysForOneClass} and
\link{clusteringTypicalDays}. The user can choose to get 
the vertices of the input data, the raw ptdf of the input data or the ptdf after 
differenciation of the input data. One of the interests of this function is to 
get the ptdf to use easily the package fbAntares
}
