% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreprocPlan.R
\name{getPreprocPlan}
\alias{getPreprocPlan}
\title{Generate a data.table with the ptdf}
\usage{
getPreprocPlan(pathPtdfMatrixFactor = "PtdfMatrixFactors.csv",
  pathPtdfMatrixConstraint = "PtdfMatrixConstraints.csv")
}
\arguments{
\item{pathPtdfMatrixFactor}{path of the file containing the ptdf of the
flowbased structure
\itemize{
 \item SESSION_ID : The Date, in numeric format : YYYYMMDD
 \item MATRIX_ID : The Period (hour in the day), between 1 and 24
 \item ROW_ID : The id of the row
 \item REMAININGAVAILABLEMARGIN : The constraint (or ram)
}}

\item{pathPtdfMatrixConstraint}{path of the file containing the constraints of 
the flowbased structure (the ram). Its columns have to be :
\itemize{
 \item SESSION_ID : The Date, in numeric format : YYYYMMDD
 \item MATRIX_ID : The Period (hour in the day), between 1 and 24
 \item ROW_ID : The id of the row
 \item BIDDINGAREA_ID : The id of the hub, we must have at least two different
 and they have to be in this selection: (23, 90, 95, 22, 20, 59, 97, 92, 41, 98, 83, 96, 99)
 \item FACTOR : The values of the ptdf by hub
}}
}
\description{
With links to the files PtdfMatrixFactors and PtdfMatrixConstraints,
this function computes a data.table containing all columns needed to keep going
on the procedure to get the distances between polyhedra.
}
\examples{

\dontrun{
library(data.table)
pathPtdfMatrixFactor = system.file(
"testdata/plan_new_version_factor_AT.rds", package = "fbClust")
pathPtdfMatrixConstraint = system.file(
"testdata/plan_new_version_constraint_AT.rds", package = "fbClust")

 PLAN <- getPreprocPlan(pathPtdfMatrixFactor = pathPtdfMatrixFactor,
 pathPtdfMatrixConstraint =  pathPtdfMatrixConstraint)
}

}
