% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDiffNotWantedPtdf.R
\name{setDiffNotWantedPtdf}
\alias{setDiffNotWantedPtdf}
\title{Compute facets equations}
\usage{
setDiffNotWantedPtdf(PLAN, hubDrop = list(NL = c("BE", "DE", "FR",
  "AT")))
}
\arguments{
\item{PLAN}{{data.table}}

\item{hubDrop}{\code{list}, list of hubs in the ptdf, with the ones which should
sustracted to the others as the names of the arrays which themself contain the ones which
be sustracted}
}
\description{
This function computes the facets equation by droping one or more hubs
and sustracting them to the others (with a named list).
}
